Add ts/api folder
- Slim down home.ts
