+++ web/ts/utilities/fetch-wrappers.ts
@@ -1,0 +1,20 @@
+/**
+ * Wrapper for Javascript's fetch function
+ * @param  {string} url URL to fetch
+ * @param  {object} default_resp Return value in case of error
+ * @return {Promise<Response>}
+ */
+export async function get(url: string, default_resp: object = []) {
+    return fetch(url)
+        .then((res) => {
+            if (!res.ok) {
+                throw Error(res.statusText);
+            }
+            return res.json();
+        })
+        .catch((err) => {
+            console.error(err);
+            return default_resp;
+        })
+        .then((o) => o);
+}
